Moved HttpRequest expr Type to automation
